fix: handle dark/light mode

+++ src/pages/UserList.jsx
@@ -11,10 +11,12 @@ import { fetchUsers, removeUser } from 'redux/user/userSlice';
 const UserList = ({ columns }) => {
   const dispatch = useDispatch();
   const { users } = useSelector((state) => state.user);
+  const { darkMode } = useSelector((state) => state.darkMode);
 
   const handleDelete = (id) => {
     const userId = id;
-    dispatch(removeUser({ userId, toast }));
+    if (window.confirm('Are you sure you wanted to delete this user'))
+      dispatch(removeUser({ userId, toast }));
   };
 
   useEffect(() => {
@@ -38,11 +40,8 @@ const UserList = ({ columns }) => {
             </Link>
             <DeleteOutline
               onClick={() => handleDelete(params.row._id)}
-              style={{
-                fontSize: '2rem',
-                color: '#ff0000',
-                cursor: 'pointer',
-              }}
+              style={{ fontSize: '2rem', cursor: 'pointer' }}
+              className={darkMode ? 'delete__dark' : 'delete__light'}
             />
           </>
         );
@@ -61,6 +60,7 @@ const UserList = ({ columns }) => {
         rowsPerPageOptions={[8]}
         checkboxSelection
         style={{ fontSize: '1.5rem' }}
+        className={darkMode && 'data__grid'}
       />
     </Container>
   );
@@ -68,17 +68,16 @@ const UserList = ({ columns }) => {
 
 const Container = styled.div`
   flex: 4;
+  background-color: ${({ theme }) => theme.bg};
 `;
-
-
 
 const EditButton = styled.button`
   border: none;
   display: block;
   padding: 0.5rem 1rem;
   text-transform: capitalize;
-  background-color: #3bb077;
-  color: var(--color-white);
+  background-color: ${({ theme }) => theme.btnEdit};
+  color: ${({ theme }) => theme.textEdit};
   border-radius: 10rem;
   cursor: pointer;
   margin-right: 1rem;
